+++ packages/moonstone/CHANGELOG.md
@@ -15,7 +15,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Fixed
 
-- Input will now show an ellipsis on the correct side based on the `value` or `placeholder`'s content.
+- `moonstone/Input` to show an ellipsis on the correct side based on the text directionality of the `value` or `placeholder` content.
 - `moonstone/VirtualList` and `moonstone/VirtualGridList` to prevent unwanted scrolling when focused with the pointer
 
 ## [1.0.0-alpha.4] - 2016-12-2